Move reporters to reporter package
